Fixes arm64 image and adds CI (#40)
Our arm64 image was busted because the mirror of alpine didn't mirror
as a multi-architecture image. This adds CI to ensure that the image
built can indeed run on the intended architecture.

NOTE: this currently uses Docker Hub image as a base, but when stable,
this repository is low traffic, so unlikely to cause an outage.